==> notebooks/1_data_collection.ipynb <==
# Olist Product Recommendation System
### Part 1 Data collection and Data Cleaning 
#### Author: Olabisi Sunmon | 10th April 2023

### Problem Statement

How can we create a customised product recommendation system using data analysis and machine learning techniques to help Olist customers discover new products and find relevant items for purchase, to boost revenue and customer purchase rates.

-------
In the notebook I will be importing the raw database from Kaggle and conducting cleaning.

----------
### Data Source 

The Olist Store provides a dataset of Brazilian ecommerce transactions, consisting of 100,000 orders from 2016 to 2018 made on various marketplaces in the country. This data consists of 8 tables with an extra dataset containing English translations.

The Data can be sourced from;https://www.kaggle.com/datasets/olistbr/brazilian-ecommerce




#Import packages 
import numpy as np
import pandas as pd
import joblib 

#### Import Files that have been downloaded from Kaggle.

# Load the Data
df_main = pd.read_csv('data/raw/olist_orders_dataset.csv')
df_customer = pd.read_csv('data/raw/olist_customers_dataset.csv')
df_geo = pd.read_csv('data/raw/olist_geolocation_dataset.csv')
df_items = pd.read_csv('data/raw/olist_order_items_dataset.csv')
df_payments = pd.read_csv('data/raw/olist_order_payments_dataset.csv')
df_reviews = pd.read_csv('data/raw/olist_order_reviews_dataset.csv')
df_products = pd.read_csv('data/raw/olist_products_dataset.csv')
df_sellers = pd.read_csv('data/raw/olist_sellers_dataset.csv')
df_product_English = pd.read_csv('data/raw/product_category_name_translation.csv')

### Data Dictionary

---

**Customer Data :**

Column Name | Data Description | Data Type |
---------------- |-----------       |-------    
`customer_id` | key to the orders dataset. Each order has a unique customer_id      | Categorical     
`customer_unique_id`    | unique identifier of a customer.      | Categorical  
`customer_zip_code` | first five digits of customer zip code | Numeric
`customer_city` | customer city name | Categorical 
`customer_state` | customer state | Categorical 


**Geo Data :**

Column Name | Data Description | Data Type |
---------------- |-----------       |-------    
`geolocation_zip_code_prefix` | latitude    | Numeric     
`geolocation_lat`    | longitude     | Numeric 
`geolocation_lng` | first five digits of customer zip code | Numeric
`geolocation_city` | city name | Categorical 
`geolocation_state` | state| Categorical 


**Items Data:**

Column Name | Data Description | Data Type |
---------------- |-----------       |-------    
`order_id` | order unique identifier   | Categorical     
`order_item_id`    | sequential number identifying number of items included in the same order.    | Numeric 
`product_id` | product unique identifier | Categorical  
`seller_id` | seller unique identifier | Categorical 
`shipping_limit_date` | Shows the seller shipping limit date for handling the order over to the logistic partner.| Categorical 
`price` | item price | Numeric 
`freight_value` |item freight value item (if an order has more than one item the freight value is splitted between items).|  Numeric 


**Payment Data :**

Column Name | Data Description | Data Type |
---------------- |-----------       |-------    
`order_id` | unique identifier of an order.    | Categorical    
`payment_sequential`    | a customer may pay an order with more than one payment method. If he does so, a sequence will be created to accommodate all     | Numeric 
`payment_type` | first five digits of customer zip code | Numeric
`payment_installments` | number of installments chosen by the customer. | Numeric
`payment_value` | transaction value.| Numeric


**Reviews Data:**

Column Name | Data Description | Data Type |
---------------- |-----------       |-------    
`review_id` | unique review identifier.   | Categorical     
`order_id`    | unique order identifier.    | Categorical 
`review_score`    | Note ranging from 1 to 5 given by the customer on a satisfaction survey.   | Numeric 
`review_comment_title` | Comment title from the review left by the customer, in Portuguese. | Categorical
`review_comment_message` | Comment message from the review left by the customer, in Portuguese. | Categorical 
`review_creation_date` | Shows the date in which the satisfaction survey was sent to the customer.| Numeric 
`review_answer_timestamp` | Shows satisfaction survey answer timestamp. | Numeric 


**Main Data :**

Column Name | Data Description | Data Type |
---------------- |-----------       |-------    
`order_id` | unique identifier of an order.    | Categorical    
`customer_id`    | key to the customer dataset. Each order has a unique customer_id.  | Categorical  
`order_status` | Reference to the order status (delivered, shipped, etc). | Categorical  
`order_purchase_timestamp` | Shows the purchase timestamp. | Numeric
`order_approved_at` | Shows the payment approval timestamp.| Numeric
`order_delivered_carrier_date` | Shows the order posting timestamp. When it was handled to the logistic partner. | Numeric
`order_delivered_customer_date` | Shows the actual order delivery date to the customer.| Numeric
`order_estimated_delivery_date` |Shows the estimated delivery date that was informed to customer at the purchase moment.| Numeric

**Product Data :**

Column Name | Data Description | Data Type |
---------------- |-----------       |-------    
`product_id` | unique product identifier.   | Categorical    
`product_category_name`    | root category of product, in Portuguese.    | Categorical   
`Product_name_lenght` | number of characters extracted from the product name. | Numeric
`product_description_lenght` | number of characters extracted from the product description. | Numeric
`product_photos_qty` | number of product published photos| Numeric
`product_weight_g` | product weight measured in grams. | Numeric
`product_length_cm` | `product_length_cm` | number of characters extracted from the product description. | Numeric
`product_height_cm` | product height measured in centimeters.| Numeric
`product_width_cm` | product width measured in centimeters.| Numeric

**Sellers Data :**

Column Name | Data Description | Data Type |
---------------- |-----------       |-------    
`seller_id` | seller unique identifier   | Categorical    
`seller_zip_code_prefix`    |first 5 digits of seller zip code    | Numeric 
`seller_city` | seller city name| Categorical
`seller_state` | seller state | Categorical

**Product English Data :**

Column Name | Data Description | Data Type |
---------------- |-----------       |-------    
`product_category_name` | category name in Portuguese   | Categorical    
`product_category_name_english`    |category name in English   | Categorical    




### Inspecting and cleaning Data 

#Function to review dataframes
def data_overview(df):

    display(df.head())
    print(df.info())
    print('')
    print("NA's present in the DF per column:")
    print('')
    print(df.isna().sum())
    display(f'There are {df.duplicated().sum()} duplicated rows')
    display(f'The shape of the DF is {df.shape}')
   

data_overview(df_main)

#### Convert to date time

df_main['order_purchase_timestamp'] = pd.to_datetime(df_main['order_purchase_timestamp'])
df_main['order_approved_at'] = pd.to_datetime(df_main['order_approved_at'])
df_main['order_delivered_carrier_date'] = pd.to_datetime(df_main['order_delivered_carrier_date'])
df_main['order_delivered_customer_date'] = pd.to_datetime(df_main['order_delivered_customer_date'])
df_main['order_estimated_delivery_date'] = pd.to_datetime(df_main['order_estimated_delivery_date'])



#Addressing Na's 
df_main.info()

df_main have a significant number of missing values, as I do not need these columns for analysis or modeling I will drop them

df_main = df_main.drop(['order_approved_at','order_delivered_carrier_date','order_delivered_customer_date','order_estimated_delivery_date'], axis=1)

data_overview(df_customer)

No cleaning necessary for df_customer.

data_overview(df_geo)

I will Delete all 261831 duplications as no new infomation is gained by keeping the duplicates.

df_geo = df_geo.drop_duplicates()

data_overview(df_items)

No cleaning necessary for df_items.

df_items['shipping_limit_date'] = pd.to_datetime(df_items['shipping_limit_date'])

data_overview(df_payments)

No cleaning necessary for df_payments.

data_overview(df_product_English)

No cleaning necessary for df_product_English.

data_overview(df_products)


df_products.isna().sum().value_counts 

There is a significant number of missing values in df_products, I will not any of these columns besides `product_category_name` for my analysis and modeling so I will drop the others.

#Dropping columns
df_products = df_products.drop(['product_name_lenght', 'product_description_lenght','product_photos_qty','product_weight_g','product_length_cm','product_height_cm','product_width_cm'], axis=1)

na = df_products[df_products['product_category_name'].isna()] # Create a DataFrame with Na values 

na.groupby('product_id')[['product_category_name']].sum() #Display

df_products['product_category_name'].nunique

All the 610 na values in the `product_category_name` column are unique products, as all 32951 `product_id`'s are unique. So I will not be able to locate the corresponding `product_category_name` in the df. Therfeore I will not be addressing the Na values in the `product_category_name` as I have no direct mapping. I will not be dropping the column as I will be using it for analysis and modelling.

data_overview(df_sellers)

data_overview(df_items)

No cleaning necessary for df_items.

data_overview(df_reviews)

df_reviews has a significant number of missing values, I will not need these columns for analysis or modeling so I will drop them.

df_reviews = df_reviews.drop(['review_comment_title', 'review_comment_message'], axis=1)

#### The dataFrames are now ready for EDA

#Export Clean Dataset 

joblib.dump(df_customer,'data/Clean/customer_clean.pkl')
joblib.dump(df_geo,'data/Clean/geo_clean.pkl')
joblib.dump(df_items,'data/Clean/items_clean.pkl')
joblib.dump(df_main,'data/Clean/main_clean.pkl')
joblib.dump(df_products,'data/Clean/products_clean.pkl')
joblib.dump(df_payments,'data/Clean/payments_clean.pkl')
joblib.dump(df_sellers,'data/Clean/sellers_clean.pkl')
joblib.dump(df_reviews,'data/Clean/reviews_clean.pkl')
joblib.dump(df_product_English,'data/Clean/product_english_clean.pkl')
