==> notebooks/3_modelling_comparing_market_basket.ipynb <==
# Olist Product Recommendation System
### Part 3 - Comparing Granularities using Market Basket 
#### Author: Olabisi Sunmon | 10th April 2023

### Problem Statement

How can we create a customised product recommendation system using data analysis and machine learning techniques to help Olist customers discover new products and find relevant items for purchase, to boost revenue and customer purchase rates.


During the exploratory data analysis I found that around 55% of customers have made only one purchase. This can pose a challenge for collaborative filtering techniques and market basket analysis as they heavily rely on user purchase history to make accurate recommendations to predict their future behavior and preferences.

### Approach
The data frame will be split into 2 by separating customers who made a single purchase (First-timers) and customers who made multiple purchases (Returners). 
The unpartitioned dataset is what will be used to create recommendations for new costumers with empty online carts and the returners dateframe will be used to build recommendation systems for returning costumers and new costumers with items in their online cart.

-------
In this notebook, I will assess the different levels of granularity for a product recommendation system for returning customers using Market Basket analysis. I will investigate the feasibility of recommending products based on 
- Product Caterogy per State
- Items in a Product Caterogy
- All products
and analyse the effectiveness of each approach.


Market basket analysis is a way to find out what items customers tend to buy together. By analysing transaction data, we can identify patterns and relationships between products. 

-------

### Procedure:

One order will be considered as one unit (a basket)
 
I will be utilising three different methods for market basket analysis due to limitations in computational capacity.

-  Hot encode (Mlxtend) 
-  TransactionEncoder (Mlxtend)
-  Manually 

#Import Package
# data manipulation
import numpy as np
import pandas as pd
import joblib
from sklearn.model_selection import StratifiedShuffleSplit

#Plotting
import matplotlib as mpl
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

#Modeling
from mlxtend.frequent_patterns import apriori, fpmax, fpgrowth, association_rules
from mlxtend.preprocessing import TransactionEncoder
from surprise import Dataset
from surprise.reader import Reader
from surprise.prediction_algorithms.matrix_factorization import SVD as FunkSVD
from surprise.model_selection import train_test_split, GridSearchCV, cross_validate
from surprise import accuracy
from surprise.accuracy import rmse
from surprise import NormalPredictor
from itertools import combinations, groupby
from collections import Counter

#Ignore futurewarnings
import warnings
warnings.filterwarnings('ignore')

#### Import data

df = joblib.load('data/Processed/df_processed.pkl')


df.shape

# Split the dataframe into 2 dataframes for first time and returners
# Creates a repeater columns of both dateframes

returners = df.groupby('customer_unique_id').filter(lambda x: len(x) > 1)
first_timers = df.groupby('customer_unique_id').filter(lambda x: len(x) == 1)


# Save datasets for easier accessing
joblib.dump(returners,'data/processed/returners_data.pkl')
joblib.dump(first_timers,'data/processed/first_timer_data.pkl')

    
print("Shape of returners dataset:", returners.shape)
print("Shape of first timer dataset:", first_timers.shape)


## Market Basket

The min_support value of 0.001 will be used in all the market basket models of this analysis, it has been chosen based on the mean support of the "all product" granularity level. Keeping the support value constant aims to maintain fairness and impartiality in the analysis, while also preserving the accuracy and validity of the results.

### Product Caterogy per State
Hot encode (Mlxtend) 

# Create basket for different state using product caterogy
basket_RR = (returners[returners['customer_state'] == 'RR'] 
          .groupby(['order_id','product_category_name_english'])['order_item_id'] 
          .sum().unstack().reset_index().fillna(0) 
          .set_index('order_id'))
basket_BA = (returners[returners['customer_state'] == 'BA'] 
          .groupby(['order_id','product_category_name_english'])['order_item_id'] 
          .sum().unstack().reset_index().fillna(0) 
          .set_index('order_id'))
basket_SP = (returners[returners['customer_state'] == 'SP'] 
          .groupby(['order_id','product_category_name_english'])['order_item_id'] 
          .sum().unstack().reset_index().fillna(0) 
          .set_index('order_id'))
basket_SP.shape

# data encoding
# defining the hot encoding function to make the data suitable 
# for the concerned libraries 
def hot_encode(x): 
    if(x<= 0): 
        return 0
    if(x>= 1): 
        return 1
basket_RR = basket_RR.applymap(hot_encode) 
basket_BA = basket_BA.applymap(hot_encode)
basket_SP = basket_SP.applymap(hot_encode) 
print('Basket SP')
basket_SP


print("For RR:")
frq_items = apriori(basket_RR, min_support = 0.001, use_colnames = True) # Building the model
rules = association_rules(frq_items, metric ="lift", min_threshold = 1)
# Collecting the inferred rules in a dataframe
rules = rules.sort_values(['confidence', 'lift'], ascending =[False, False]) 
display(rules.head())

#Sorting values for confidence and lift is important because it allows us to identify the most meaningful and relevant association rules in our dataset.

print("For BA:")
frq_items = apriori(basket_BA, min_support = 0.001, use_colnames = True) 
rules = association_rules(frq_items, metric ="lift", min_threshold = 1) 
rules = rules.sort_values(['confidence', 'lift'], ascending =[False, False]) 
display(rules.head())

print("For SP:")
frq_items = apriori(basket_SP, min_support = 0.001, use_colnames = True) 
rules = association_rules(frq_items, metric ="lift", min_threshold = 1) 
rules = rules.sort_values(['confidence', 'lift'], ascending =[False, False]) 
display(rules.head())

#### Analysis

For BA:

If a customer buys from the fashion_sport department, then they are very likely to buy from fashio_female_clothing.The higher the 'lift', the higher likelihood that 'consequents' will be bought after 'antecedents'. The lift for this combination is very at 456.

For RR: 

The results of the market basket analysis did not generate any recommendations for the state "RR" , indicating that developing a recommendation system at the level of individual states may not be a viable strategy at this moment in time.



#### Method
The Hot Encode method and Mlxtend library were used successfully in the analysis without anycomputational challenges. Although Hot Encode method produces a sparse matrix of integers (which takes up more memory than boolean values), Mlxtend's Apriori algorithm was still able to run efficiently. This was likely due to the small tables due to the state partition and the high level grouping of product caterogy.

### Items in a product Caterogy
TransactionEncoder (Mlxtend)

# Create baskets for items in product caterogy
my_basket_drinks = returners[returners['product_category_name_english'] == 'drinks'].groupby('order_id')['product_id'].apply(list)
my_basket_flowers = returners[returners['product_category_name_english'] == 'flowers'].groupby('order_id')['product_id'].apply(list)



#Create Boolen matrix

te = TransactionEncoder()
drinks = te.fit_transform(my_basket_drinks)
drinks_df = pd.DataFrame(drinks, columns=te.columns_)

flowers = te.fit_transform(my_basket_flowers)
flowers_df = pd.DataFrame(flowers, columns=te.columns_)

print('Flowers')
flowers_df

print("For Drinks:")
# Building the model 
frq_items = apriori(drinks_df,  min_support = 0.001 ,use_colnames = True) 
# Collecting the inferred rules in a dataframe 
rules_d = association_rules(frq_items, metric ="lift", min_threshold = 1)
rules_d = rules_d.sort_values(['confidence', 'lift'], ascending =[False, False]) 
display(rules_d.head())

print("For Flowers:")
# Building the model 
frq_items = apriori(flowers_df,  min_support = 0.001 ,use_colnames = True) 
# Collecting the inferred rules in a dataframe 
rules_f = association_rules(frq_items, metric ="lift", min_threshold = 1)
rules_f = rules_f.sort_values(['confidence', 'lift'], ascending =[False, False]) 
display(rules_f.head())



#### Analysis
For Drinks:

Based on the market basket analysis, there appears to be a strong positive association between the items '262855d4dd5b0d39f786a3c86c285c6a' and '7d873ae7ba4e637167c1a8d544bd6af8', suggesting that customers who purchase one of these items are likely to be interested in the other.

With a lift value of 68, this association is particularly strong. Recommending a product bundle containing both items could potentially be an effective strategy for encouraging customers to purchase them together. Therefore, it may be beneficial for Olist to consider this as it could help to increase sales and improve customer satisfaction by providing customers with a convenient and attractive offer that meets their needs and preferences.

For Flowers:

The market basket analysis did not create any recommendations for the "flowers" product category, suggesting that implementing a recommendation system at the level of 'product caterogy may be premature. This granuality should been revisted when olist has a larger amount of order history data.


#### Method

To avoid overburdening the computational resources for this model, the TransactionEncoder property was utilised. There are many items in some product categories, which will result in large sparse matrices. TransactionEncoder() was used to create a boolean sparse matrix that occupies much less memory compared to an integer sparse matrix. Previously, an integer sparse matrix was used, but it was not able to undergo the apriori transformation due to limited computational power for this level of granularity.

### All Products
Manual

all_products = df.set_index('order_id')['product_id']
all_products.head()

stats = all_products.value_counts().to_frame('frequency')
stats.head()

# calc support
stats['support'] = stats/ len(set(all_products.index))*100
stats.head()
     

mean = np.mean(stats['support'])
print(f'mean: {mean}')

max = np.max(stats['support']) 
print(f'max: {max}')

min = np.min(stats['support'])
print(f'min: {min}')

median = np.median(stats['support'])
print(f'median: {median}')


# min_support level
min_support = 0.001 #using the median
product_over_support = stats[stats['support'] >= min_support].index
orders_over_support = all_products[all_products.isin(product_over_support)]
orders_over_support
     

#filter out order that only had 1 item order on the order_id
order_count = all_products.index.value_counts()
order_count.head()


# create a func that will generate our itemsets and send our new order dataset through the generator
def generator(order):
    order = order.reset_index().values
    for order_id, product in groupby(order, lambda x: x[0]):
        product_list = [item[1] for item in product]
        for item_pair in combinations(product_list, 2):
            yield item_pair

itempair_gen = generator(orders_over_support)
itempair_gen

itempair = pd.Series(Counter(itempair_gen)).to_frame('freqAC')
itempair['supportAC'] = itempair['freqAC'] / len(orders_over_support) *100
itempair = itempair[itempair['supportAC'] >= min_support]
itempair.head()

# create table for association rules and compute relevant metrics
itempair = itempair.reset_index().rename(columns={'level_0' : 'antecedents', 'level_1': 'consequents'})
itempair = (itempair
     .merge(stats.rename(columns={'freq': 'freqA', 'support': 'antecedent support'}), left_on='antecedents', right_index=True)
     .merge(stats.rename(columns={'freq': 'freqC', 'support': 'consequents support'}), left_on='consequents', right_index=True))

itempair['confidenceAtoC'] = itempair['supportAC'] / itempair['antecedent support']
itempair['confidenceCtoA'] = itempair['supportAC'] / itempair['consequents support']
itempair['lift'] = itempair['supportAC'] / (itempair['antecedent support'] * itempair['consequents support'])
itempair = itempair[['antecedents', 'consequents','antecedent support', 'consequents support', 'confidenceAtoC','lift']]   

itempair.head()

# to find items that have confidence > 0.0001
conf = itempair
market_basket = conf[(conf.confidenceAtoC > 0.0001)]
market_basket.set_index('antecedents', inplace = True)
market_basket.reset_index(inplace=True)
market_basket = market_basket.sort_values('lift', ascending = False)

market_basket

#### Analysis

Based on the market basket analysis, there appears to be a strong positive association between the items '270516a3f41dc035aa87d220228f844c' and '05b515fdc76e888aada3c6d66c201dff', suggesting that customers who purchase one of these items are likely to be interested in the other.

#### Method

The dataset has not been further partitioned, and due to computational limitations when using Mlxtend library, a manual process was used for this example.
Source; https://pythondata.com/market-basket-analysis-with-python-and-pandas/

Based on the results obtained thus far, I plan to develop a recommendation system function that utilises the "all product" level of granularity. This approach may be more effective than selecting data at a specific level of granularity, as the latter has not produced optimal results thus far, as seen in the earlier analysis.

#Recommend product based on association rule what item is in the basket ,market basket tableand  metric of interest
# Input basket
mybasket = ['d1c427060a0f73f6b889a5c7c61f2ac4']

# metric
metric = "lift"

def product_recs(basket, df, metric):
   
    # Randomly select an item from the basket
    random_item = np.random.choice(basket, 1)[0]
    print(f"Since you like {random_item}, you might like: ")

    # Find rules where the item is in the antecedent
    rule_filter = df["antecedents"].apply(lambda x: x) == random_item

    # Filter the dataframe using rule_filter and sort by the selected metric
    filtered = df[rule_filter].sort_values(by=metric, ascending=False)

    # Randomly return one of the top 10 items from the filtered dataframe
    recommendation = filtered.head(10).sample()["consequents"]

    return recommendation


product_recs(mybasket, market_basket, metric)


To ensure that the best-fitted recommendations are given, I suggest that Olist uses the product recommedation system (product_recs) to generate recommendations for customers when they are at the online checkout section.This should presented as bundle deal to encourage the costumer to purchase the recommendation. As Market basket analysis is most effective when applied to the final combination of products.This recommendation is suitable for both new and existing customers, provided they have items in their online shopping cart.
